% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/juan.R
\name{gitjuan}
\alias{gitjuan}
\title{An homage to our favorite student, Juan}
\usage{
gitjuan(n = 1)
}
\arguments{
\item{n}{the number of letters you want to draw}
}
\value{
a vector of n randomly selected letters, with replacement
}
\description{
Allows for random sampling of letters from the built-in letters object
}
