\name{juanmoretime-package}
\alias{juanmoretime-package}
\alias{juanmoretime}
\docType{package}
\title{
\packageTitle{juanmoretime}
}
\description{
\packageDescription{juanmoretime}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{juanmoretime}
\packageIndices{juanmoretime}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{juanmoretime}

Maintainer: \packageMaintainer{juanmoretime}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
